requirements.txt
# Innovation Intelligence System - Python Dependencies

# LangChain Core
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.1

# OpenAI API Client (required by langchain-openai)
openai>=1.0.0

# Document Processing
pypdf>=3.17.0

# Data Handling
pyyaml>=6.0

# Templating
jinja2>=3.1.3

# Environment Management
python-dotenv>=1.0.0

# Data Validation (optional)
pydantic>=2.5.0
